FEATURE remade naming contexts
also added a feature to allow for multiple entry databases... when using
this feature is required to handle an special deployment case (with
multiple AS instances). This allows for better granularity of the
service and work load balancing but requires a lot more server features.
Since JNDI names have been modified an updated version of the AS
(patched for custom performance) should be used for testing.